# src/OnDeploy/docs/swagger.yaml
basePath: /api/v1
definitions:
  models.ComRabbitPermission:
    properties:
      address:
        type: string
      user:
        type: string
      vhost:
        type: string
    type: object
  models.CommonRabbitUser:
    properties:
      address:
        type: string
      user:
        type: string
    type: object
  models.Err:
    $ref: '#/definitions/models.Res'
...
  models.NewRabbitPermission:
    properties:
      address:
        type: string
      configure:
        type: string
      read:
        type: string
      user:
        type: string
      vhost:
        type: string
      write:
        type: string
    type: object
  models.NewRabbitUser:
    properties:
      address:
        type: string
      pass:
        type: string
      user:
        type: string
    type: object
  models.NewRabbitVhost:
    properties:
      address:
        type: string
      tracing:
        type: boolean
      vhost:
        type: string
    type: object
  models.Res:
    properties:
      code:
        type: integer
      msg:
        type: string
    type: object
  models.ServerDetail:
    properties:
      address:
        type: string
      name:
        type: string
      port:
        description: |-
          Username string `json:"username,omitempty"`
          Password string `json:"password,omitempty"`
        type: integer
    type: object
  models.ServersDetail:
    items:
      $ref: '#/definitions/models.ServerDetail'
    type: array
info:
  contact: {}
  description: 应用交付部署服务API
  license: {}
  title: 应用交付部署服务
  version: "1.0"
paths:
...
  /app/rabbitmq/install:
    post:
      consumes:
      - application/json
      description: RabbitMQ服务安装
      parameters:
      - description: server
        in: body
        name: server
        required: true
        schema:
          $ref: '#/definitions/models.ServerDetail'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Res'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Err'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.Err'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Err'
      security:
      - basic: []
      summary: RabbitMQ服务安装
      tags:
      - RabbitMQ服务
  /app/rabbitmq/permission/add:
    put:
      consumes:
      - application/json
      description: RabbitMQ权限添加
      parameters:
      - description: permission
        in: body
        name: permission
        required: true
        schema:
          $ref: '#/definitions/models.NewRabbitPermission'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Res'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Err'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Err'
      summary: RabbitMQ权限添加
      tags:
      - RabbitMQ服务
  /app/rabbitmq/permission/del:
    delete:
      consumes:
      - application/json
      description: RabbitMQ权限删除
      parameters:
      - description: permission
        in: body
        name: permission
        required: true
        schema:
          $ref: '#/definitions/models.ComRabbitPermission'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Res'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Err'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Err'
      summary: RabbitMQ权限删除
      tags:
      - RabbitMQ服务
  /app/rabbitmq/permission/lst/{address}:
    get:
      consumes:
      - application/json
      description: RabbitMQ权限列表
      parameters:
      - description: address
        in: path
        name: address
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Res'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Err'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Err'
      summary: RabbitMQ权限列表
      tags:
      - RabbitMQ服务
  /app/rabbitmq/user/add:
    put:
      consumes:
      - application/json
      description: RabbitMQ用户添加
      parameters:
      - description: user
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/models.NewRabbitUser'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Res'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Err'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Err'
      summary: RabbitMQ用户添加
      tags:
      - RabbitMQ服务
  /app/rabbitmq/user/del:
    delete:
      consumes:
      - application/json
      description: RabbitMQ用户删除
      parameters:
      - description: user
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/models.CommonRabbitUser'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Res'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Err'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Err'
      summary: RabbitMQ用户删除
      tags:
      - RabbitMQ服务
  /app/rabbitmq/user/lst/{address}:
    get:
      consumes:
      - application/json
      description: RabbitMQ用户列表
      parameters:
      - description: address
        in: path
        name: address
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Res'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Err'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Err'
      summary: RabbitMQ用户列表
      tags:
      - RabbitMQ服务
  /app/rabbitmq/vhost/add:
    put:
      consumes:
      - application/json
      description: RabbitMQ虚拟机添加
      parameters:
      - description: vhost
        in: body
        name: vhost
        required: true
        schema:
          $ref: '#/definitions/models.NewRabbitVhost'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Res'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Err'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Err'
      summary: RabbitMQ虚拟机添加
      tags:
      - RabbitMQ服务
  /app/rabbitmq/vhost/del:
    delete:
      consumes:
      - application/json
      description: RabbitMQ虚拟机删除
      parameters:
      - description: vhost
        in: body
        name: vhost
        required: true
        schema:
          $ref: '#/definitions/models.NewRabbitVhost'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Res'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Err'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Err'
      summary: RabbitMQ虚拟机删除
      tags:
      - RabbitMQ服务
  /app/rabbitmq/vhost/lst/{address}:
    get:
      consumes:
      - application/json
      description: RabbitMQ虚拟机列表
      parameters:
      - description: address
        in: path
        name: address
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Res'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Err'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Err'
      summary: RabbitMQ虚拟机列表
      tags:
      - RabbitMQ服务
  /server/init:
    post:
      consumes:
      - application/json
      description: 单台服务器初始化
      parameters:
      - description: server
        in: body
        name: server
        required: true
        schema:
          $ref: '#/definitions/models.ServerDetail'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Res'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Err'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.Err'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Err'
      security:
      - basic: []
      summary: 单台服务器初始化
      tags:
      - 服务器
  /servers/init:
    post:
      consumes:
      - application/json
      description: 批量服务器初始化
      parameters:
      - description: servers
        in: body
        name: server
        required: true
        schema:
          $ref: '#/definitions/models.ServersDetail'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Res'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Err'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.Err'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Err'
      security:
      - basic: []
      summary: 批量服务器初始化
      tags:
      - 服务器
schemes:
- http
- https
securityDefinitions:
  basic:
    type: basic
swagger: "2.0"